--- lib/presentation/screen/profile/profile_screen.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileScreen extends StatelessWidget {
-  const ProfileScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(body: const Center(child: Text('Profile Screen')));
-  }
-}
